+++ modules/control_system/functions/d2c.m
@@ -1,47 +1,52 @@
-% Transform discrete transfer function or state space model to continuous time
-% transfer function or state space model 
-% Input: Gd, sysd
-% Example 1: G = mc.d2c(Gd)
-% Example 2: sys = mc.d2c(sysd)
-% Author: Daniel Mårtensson, September 2017
-
-function [model] = d2c(varargin)
-    if(isempty(varargin{1}))
-      error ('Missing model')
-    end
-    
-    % State space
-    if(strcmp(varargin{1}.type, 'SS' ))
-      % Get info
-      A = varargin{1}.A;
-      B = varargin{1}.B;
-      C = varargin{1}.C;
-      D = varargin{1}.D;
-      delay = varargin{1}.delay;
-      sampleTime = varargin{1}.sampleTime;
-      % Compute sizes
-      a1 = size(A,2) + size(B,2) - size(A,1);
-      b1 = size(A,2);
-      a2 = size(A,2) + size(B,2) - size(B,1);
-      b2 = size(B,2);
-      % Compute square matrix
-      M = [A B; zeros(a1, b1)  eye(a2, b2)];
-      M = round(logm(M)*1/sampleTime*1/1e-6)*1e-6; % Very important to remove small numbers!!
-      M(abs(M) < 1e-6) = 0; % Very importat to turn -0 to 0.
-      A = M(1:size(A,1), 1:size(A,2));
-      B = M(1:size(B,1), (size(A,2) + 1):(size(A,2) + size(B,2)));
-      % Return model
-      model = mc.ss(delay, A, B, C, D);
-      % Don't forget to add sample time to zero again
-      model.sampleTime = 0;
-    elseif(strcmp(varargin{1}.type, 'TF' ))
-      % Turn TF to SS
-      sysd = mc.tf2ss(varargin{1}, 'OCF');
-      % Call d2c2
-      G = mc.d2c(sysd);
-      % SS to SS
-      model = mc.ss2tf(G);
-    else
-      error('No state space model or transfer function')
+%=============================================================================
+% Copyright (c) 2017 September Daniel Mårtensson (Swedish Embedded Control Systems Toolbox)
+% Copyright (c) 2023-present Allan CORNET (Nelson)
+%=============================================================================
+% This file is part of the Nelson.
+%=============================================================================
+% LICENCE_BLOCK_BEGIN
+% SPDX-License-Identifier: LGPL-3.0-or-later
+% LICENCE_BLOCK_END
+%=============================================================================
+function varargout = d2c(varargin)
+  % [a, b] = d2c(phi, gamma, t)
+  
+  narginchk(3,3);
+  nargoutchk(0, 2);
+  
+  phi = varargin{1};
+  gamma = varargin{2};
+  T = varargin{3};
+  
+  [msg, phi, gamma] = abcdchk(phi, gamma);
+  if ~isempty(msg)
+    error(msg);
+  end
+  
+  [m, n] = size(phi);
+  [m, nb] = size(gamma);
+  
+  if m == 1 && phi == 1
+    a = 0;
+    b = gamma / T;
+    return
+  end
+  
+  b = zeros(m, nb);
+  nz = 0;
+  nonzero = [];
+  for i=1:nb
+    if (any(gamma(:,i ) ~= 0)) 
+      nonzero = [nonzero, i];
+      nz = nz + 1;
     end
   end
+  
+  m = [[phi, gamma(:, nonzero)]; zeros(nz, n), eye(nz)];
+  [s, dummy] = logm(m);
+  s = real(s / T);
+  a = s(1:n, 1:n);
+  if length(b)
+    b(:, nonzero) = s(1:n, n + 1 : n + nz);
+  end
+end